Removed sudo functionality, using umask instead.
This commit removes the functionality and necessity for
sudo calls between the openproject and git users.

As gitolite suggests, we now access the repositories through
the git group and a umask of 0770.

For more information on the gitolite config, see:
http://gitolite.com/gitolite/g2/rc.html

This commit also adds a few changes to the configuration:

* delayed_job can be disabled through the config and is disabled by
default
* The relative git storage dir has been replaced with an absolute path
setting for "Gitolite repositories base path".
* Repository::Git no longer store the url as a relative path from the
git home directory, but only the relative path below the base_path
* Added an output of the repository names in the config test